+++ EFT_S9/src/main/java/com/duoc/eft_s9/exceptions/ArriendoPersistenciaException.java
@@ -1,0 +1,7 @@
+package com.duoc.eft_s9.exceptions;
+
+public class ArriendoPersistenciaException extends Exception {
+    public ArriendoPersistenciaException(String mensaje) {
+        super(mensaje);
+    }
+}