+++ EFT_S9/src/main/java/com/duoc/eft_s9/exceptions/DiasArriendoInvalidosException.java
@@ -1,0 +1,7 @@
+package com.duoc.eft_s9.exceptions;
+
+public class DiasArriendoInvalidosException extends Exception {
+    public DiasArriendoInvalidosException(String mensaje) {
+        super(mensaje);
+    }
+}