+++ EFT_S9/src/main/java/com/duoc/eft_s9/exceptions/HistorialLecturaException.java
@@ -1,0 +1,7 @@
+package com.duoc.eft_s9.exceptions;
+
+public class HistorialLecturaException extends Exception {
+    public HistorialLecturaException(String mensaje) {
+        super(mensaje);
+    }
+}