+++ EFT_S9/src/main/java/com/duoc/eft_s9/exceptions/VehiculoNoDisponibleException.java
@@ -1,0 +1,7 @@
+package com.duoc.eft_s9.exceptions;
+
+public class VehiculoNoDisponibleException extends Exception {
+    public VehiculoNoDisponibleException(String mensaje) {
+        super(mensaje);
+    }
+}